feat!: Add method for fetching extended card
# Description

This change adds support for a new JSON-RPC method `agent/getAuthenticatedExtendedCard` for fetching authenticated extended card.

This is only applicable if `AgentCard.supportsAuthenticatedExtendedCard` bit is set to `true` that allows fetching authenticated context specific agent card.

In the previous versions, the extended card was served via `{AgentCard.url}/../agent/authenticatedExtendedCard` endpoint. This endpoint is being removed in favor of the JSON-RPC method. This endpoint will be deprecated in the SDKs and will be removed in a future release.

Release-As: 0.3.0